+++ packages/broker/src/plugins/storage/dataQueryEndpoint.ts
@@ -146,7 +146,7 @@ const handleFrom = (
 ) => {
     metrics.resendFromQueriesPerSecond.record(1)
     const fromTimestamp = parseIntIfExists(req.query.fromTimestamp)
-    const fromSequenceNumber = parseIntIfExists(req.query.fromSequenceNumber) || MIN_SEQUENCE_NUMBER_VALUE
+    const fromSequenceNumber = parseIntIfExists(req.query.fromSequenceNumber) ?? MIN_SEQUENCE_NUMBER_VALUE
     const { publisherId } = req.query
     if (fromTimestamp === undefined) {
         sendError('Query parameter "fromTimestamp" required.', res)
@@ -179,8 +179,8 @@ const handleRange = (
     metrics.resendRangeQueriesPerSecond.record(1)
     const fromTimestamp = parseIntIfExists(req.query.fromTimestamp)
     const toTimestamp = parseIntIfExists(req.query.toTimestamp)
-    const fromSequenceNumber = parseIntIfExists(req.query.fromSequenceNumber) || MIN_SEQUENCE_NUMBER_VALUE
-    const toSequenceNumber = parseIntIfExists(req.query.toSequenceNumber) || MAX_SEQUENCE_NUMBER_VALUE
+    const fromSequenceNumber = parseIntIfExists(req.query.fromSequenceNumber) ?? MIN_SEQUENCE_NUMBER_VALUE
+    const toSequenceNumber = parseIntIfExists(req.query.toSequenceNumber) ?? MAX_SEQUENCE_NUMBER_VALUE
     const { publisherId, msgChainId } = req.query
     if (req.query.fromOffset !== undefined || req.query.toOffset !== undefined) {
         sendError('Query parameters "fromOffset" and "toOffset" are no longer supported. Please use "fromTimestamp" and "toTimestamp".', res)