+++ packages/client/test/unit/OrderMessages.test.ts
@@ -27,7 +27,8 @@ const createTransform = (resends: Pick<Resends, 'resend'>, config = CONFIG) => {
         config as any,
         resends as any,
         STREAM_PART_ID,
-        mockLoggerFactory()
+        mockLoggerFactory(),
+        undefined as any
     ).transform()
 }
 
@@ -86,7 +87,8 @@ describe('OrderMessages', () => {
                 publisherId: PUBLISHER_ID,
                 msgChainId: MSG_CHAIN_ID,
                 raw: true
-            }
+            },
+            undefined
         )
     })
 
@@ -113,7 +115,8 @@ describe('OrderMessages', () => {
                 publisherId: PUBLISHER_ID,
                 msgChainId: MSG_CHAIN_ID,
                 raw: true
-            }
+            },
+            undefined
         )
     })
 
@@ -143,7 +146,8 @@ describe('OrderMessages', () => {
                 publisherId: PUBLISHER_ID,
                 msgChainId: MSG_CHAIN_ID,
                 raw: true
-            }
+            },
+            undefined
         )
         expect(resends.resend).toHaveBeenNthCalledWith(2,
             STREAM_PART_ID,
@@ -159,7 +163,8 @@ describe('OrderMessages', () => {
                 publisherId: PUBLISHER_ID,
                 msgChainId: MSG_CHAIN_ID,
                 raw: true
-            }
+            },
+            undefined
         )
     })